[minor] `private[ox]` method moved too to `collections.scala` file
