+++ core/src/main/scala/ox/collections.scala
@@ -3,6 +3,24 @@ package ox
 import scala.annotation.unused
 import scala.collection.IterableFactory
 
+import java.util.concurrent.Semaphore
+
+private[ox] def commonPar[I, O, C[E] <: Iterable[E], FO](parallelism: Int, iterable: => C[I], transform: I => O, handleOutputs: Seq[O] => FO): FO =
+  val s = Semaphore(parallelism)
+
+  supervised {
+    val forks = iterable.map { elem =>
+      s.acquire()
+      fork {
+        val o = transform(elem)
+        s.release()
+        o
+      }
+    }
+    val outputs = forks.toSeq.map(f => f.join())
+    handleOutputs(outputs)
+  }
+
 /** Runs partial function in parallel on each element of `iterable` for which the partial function is defined.
  * If function is not defined for an element such element is skipped.
  * Using not more than `parallelism` forks concurrently.
